# [wpimath] Use SDA algorithm instead of SSCA for DARE solver