Both seem to work, but the SDA algorithm is specifically recommended for solving DAREs as opposed to P-DAREs.

The QR decomposition was replaced with a partial pivoting LU decomposition at the recommendation of section 2.4 of the paper.