@calcmogul 2023-08-09
Old perf numbers:
```
[==========] Running 8 tests from 1 test suite.
[----------] Global test environment set-up.
[----------] 8 tests from DARETest
[ RUN      ] DARETest.NonInvertibleA_ABQR
0.142 ms
[       OK ] DARETest.NonInvertibleA_ABQR (0 ms)
[ RUN      ] DARETest.NonInvertibleA_ABQRN
0.039 ms
[       OK ] DARETest.NonInvertibleA_ABQRN (0 ms)
[ RUN      ] DARETest.InvertibleA_ABQR
0.049 ms
[       OK ] DARETest.InvertibleA_ABQR (0 ms)
[ RUN      ] DARETest.InvertibleA_ABQRN
0.037 ms
[       OK ] DARETest.InvertibleA_ABQRN (0 ms)
[ RUN      ] DARETest.FirstGeneralizedEigenvalueOfSTIsStable_ABQR
0.014 ms
[       OK ] DARETest.FirstGeneralizedEigenvalueOfSTIsStable_ABQR (0 ms)
[ RUN      ] DARETest.FirstGeneralizedEigenvalueOfSTIsStable_ABQRN
0.012 ms
[       OK ] DARETest.FirstGeneralizedEigenvalueOfSTIsStable_ABQRN (0 ms)
[ RUN      ] DARETest.IdentitySystem_ABQR
0.024 ms
[       OK ] DARETest.IdentitySystem_ABQR (0 ms)
[ RUN      ] DARETest.IdentitySystem_ABQRN
0.012 ms
[       OK ] DARETest.IdentitySystem_ABQRN (0 ms)
[----------] 8 tests from DARETest (0 ms total)
```
New perf numbers:
```
[==========] Running 8 tests from 1 test suite.
[----------] Global test environment set-up.
[----------] 8 tests from DARETest
[ RUN      ] DARETest.NonInvertibleA_ABQR
0.1 ms
[       OK ] DARETest.NonInvertibleA_ABQR (0 ms)
[ RUN      ] DARETest.NonInvertibleA_ABQRN
0.024 ms
[       OK ] DARETest.NonInvertibleA_ABQRN (0 ms)
[ RUN      ] DARETest.InvertibleA_ABQR
0.035 ms
[       OK ] DARETest.InvertibleA_ABQR (0 ms)
[ RUN      ] DARETest.InvertibleA_ABQRN
0.021 ms
[       OK ] DARETest.InvertibleA_ABQRN (0 ms)
[ RUN      ] DARETest.FirstGeneralizedEigenvalueOfSTIsStable_ABQR
0.01 ms
[       OK ] DARETest.FirstGeneralizedEigenvalueOfSTIsStable_ABQR (0 ms)
[ RUN      ] DARETest.FirstGeneralizedEigenvalueOfSTIsStable_ABQRN
0.009 ms
[       OK ] DARETest.FirstGeneralizedEigenvalueOfSTIsStable_ABQRN (0 ms)
[ RUN      ] DARETest.IdentitySystem_ABQR
0.018 ms
[       OK ] DARETest.IdentitySystem_ABQR (0 ms)
[ RUN      ] DARETest.IdentitySystem_ABQRN
0.009 ms
[       OK ] DARETest.IdentitySystem_ABQRN (0 ms)
[----------] 8 tests from DARETest (0 ms total)
```