@calcmogul 2023-08-09
50-60% of the DARE runtime seems to be in precondition checking. We may want to consider a "trust me" version of the function for cases where the DARE is solved repeatedly.