@calcmogul 2023-08-12
Here's some benchmarks for different DARE solvers: https://github.com/calcmogul/DAREBench/
```
Running ./build/DAREBench
Run on (8 X 3600.24 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 256 KiB (x4)
  L3 Unified 6144 KiB (x1)
Load Average: 0.92, 1.00, 1.00
-------------------------------------------------------
Benchmark             Time             CPU   Iterations
-------------------------------------------------------
DARE_WPIMath      28064 ns        28045 ns        25089
DARE_SLICOT       41198 ns        41169 ns        16995
DARE_Drake        91757 ns        91678 ns         7658
```

Disabling the expensive precondition checks ((A, B) stabilizability and (A, C) detectability) with NDEBUG gives a 30% performance improvement, but the solver will hang if those preconditions are violated.