@calcmogul 2023-08-13
Some performance numbers for a roboRIO:
```
WPIMath: 1.048 ms
WPIMath internal: 0.54 ms
Drake: 2.505 ms
```
I couldn't get BLAS and LAPACK compiled for the roboRIO, so SLICOT is omitted.